requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==0.27.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-decouple==3.8